django/website/requirements/req_base.txt
bpython
Pillow==4.*
Django==2.2.*
django-adminactions==1.*
django-extensions==1.*
django-cors-headers==2.4.*
wagtail==2.6.*
wagtail_grapple
